Refactor and fix font asset path
